Set angle of robot immediately, instead of when close to ball
The angle of the robot was only set when the robot was close to the ball. This led to the robot first driving to the ball, and only then rotating into the right direction. This is inefficient and could lead to the robot hitting the ball while rotating instead of receiving it with the dribbler. By setting it immediately when getBall is run, the robot rotates earlier and is better able to retrieve the ball.