Change condition when kick is finished
Instead of judging whether the ball ik kicked based on its velocity, check whether the robot still has the ball. If not, the ball has been kicked. With the previous condition, the ball would think it had not kicked, even though it had, because the velocity of the ball was still very small. However, in KickAtPos, it was checked whether the robot has the ball using ->hasBall(), which checked based on distance and angle to the ball (and ballsensor). ->HasBall would think the robot lost the ball before Kick realised the ball had been kicked. This caused the KickAtPos to fail, as the robot lost the ball before kicking. Now, these checks are the same, and Kick realises the ball has been kicked correctly, avoiding this issue.