Do not get too close to ball if robot is not facing forward
In GetBall, the robot would rotate and move towards the ball simultaneously. However, if the ball is close to the robot, it would not be done rotating before already getting to the ball, which caused it to sometimes hit the ball with the side of the robot. By ensuring that the robot is facing the right direction before getting into dribbling range, it is made sure that the robot gets the ball with the dribbler and not hit it with the side.