Re-enable and fix limitAngularVel
limitAngularVel was missing a check to see if we are playing as right, which caused the calculated angle to be off. When calculating the angular velocity later on, the angle PID controller would see a very large error, as the target angle was almost pi away from the current angle (as the target angle was not adjusted for playing on the right), which caused the PID to output a very large angular velocity, causing the robot to move erratically.