Adjust HAS_BALL_ANGLE_ERROR_MARGIN from 0.1 to 0.12
The error margin was too small, which caused the robot to think it did not have the ball, when in fact the ball was touching the dribbler in the simulator. This led to the robot trying to get the ball by rotating, but the ball would simply move along with the robot, as it was touching the dribbler. This caused the robot to spin around constantly. This bigger margin avoids this issue and seems to be more representative of how the simulator works. This needs additional testing in real life to ensure this does not cause any issues on the field.